+++ src/main/resources/db/migration/20220124163145336__assign_order_create_right_with_requisition_approve_roles.sql
@@ -1,0 +1,5 @@
+INSERT INTO referencedata.role_rights (roleid, rightid)
+SELECT roleid, (SELECT id FROM referencedata.rights WHERE name = 'ORDER_CREATE')
+FROM referencedata.role_rights
+WHERE rightid = (SELECT id FROM referencedata.rights WHERE name = 'REQUISITION_APPROVE');
+