+++ src/main/resources/db/migration/20220126221943207__change_order_create_right_type_to_supervision.sql
@@ -1,0 +1,3 @@
+UPDATE referencedata.rights
+SET type = 'SUPERVISION'
+WHERE name = 'ORDER_CREATE';