+++ src/main/resources/db/migration/20220203105139390__add_cce_inventory_transfer_right.sql
@@ -1,0 +1,1 @@
+INSERT INTO referencedata.rights (id, name, type) VALUES ('8c748f30-feaf-4900-bcf6-b67bc28cbcee', 'CCE_INVENTORY_TRANSFER', 'SUPERVISION');