+++ CHANGELOG.md
@@ -6,6 +6,7 @@ New functionality added in a backwards-compatible manner:
 * [OLMIS-7444](https://openlmis.atlassian.net/browse/OLMIS-7444): Add LotValidator extension
 * [OLMIS-7452](https://openlmis.atlassian.net/browse/OLMIS-7452): Add CREATE_ORDER right and assign to users with REQUISITION_APPROVE 
 * [OLMIS-6546](https://openlmis.atlassian.net/browse/OLMIS-6546): Add product code unique validation rule to OrderableValidator
+* [OLMIS-7501](https://openlmis.atlassian.net/browse/OLMIS-7501): Add CCE_INVENTORY_TRANSFER right
 
 Breaking changes:
 * [OLMIS-7472](https://openlmis.atlassian.net/browse/OLMIS-7472): Use postgres v12