+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Bux fixes:
 
 Improvements:
 * [OLMIS-7568](https://openlmis.atlassian.net/browse/OLMIS-7568): Use openlmis/dev:7
+* [OLMIS-7568](https://openlmis.atlassian.net/browse/OLMIS-7568): Use openlmis/service-base:6
 
 15.2.2 / 2021-10-29
 ==================